test: consolidate root-level test cases into wnc_test.go
